-fix Some cutom exceptoins created and updated

+++ Bookory/src/Bookory.Business/Services/Implementations/BasketItemService.cs
@@ -18,25 +18,23 @@ public class BasketItemService : IBasketItemService
         _basketItemRepository = basketItemRepository;
     }
 
- 
-
     public async Task<BasketItem> GetBasketItemByBookIdAsync(Guid id)
     {
         var basketItem = await _basketItemRepository.GetSingleAsync(bi => bi.BookId == id);
         if (basketItem is null)
-            throw new BookNotFoundException($"Book does not exist in basket");
+            throw new BookNotFoundException("The book does not exist in the basket.");
 
         return basketItem;
     }
 
     public async Task<List<BasketItem>> GetBasketItemBySessionIdAsync(Guid id)
     {
-        List<BasketItem> basketItems =await _basketItemRepository.GetFiltered(bi => bi.SessionId == id).ToListAsync();
+        List<BasketItem> basketItems = await _basketItemRepository.GetFiltered(bi => bi.SessionId == id).ToListAsync();
 
         if (basketItems.Count == 0)
-            throw new BasketItemNotFoundException("Basket Item Not Found");
+            throw new BasketItemNotFoundException("No basket items found.");
 
-        return basketItems; 
+        return basketItems;
     }
 
     public async Task<BasketItem> GetExistingBasketItemAsync(Guid userSessionId, Guid bookId)
@@ -49,16 +47,21 @@ public class BasketItemService : IBasketItemService
         _basketItemRepository.Update(basketItem);
         await _basketItemRepository.SaveAsync();
 
-        return new ResponseDto((int)HttpStatusCode.OK, "Basket item is updated");
+        return new ResponseDto((int)HttpStatusCode.OK, "Basket item has been updated.");
     }
 
     public async Task<ResponseDto> DeleteBasketItemAsync(Guid id)
     {
         var basketItem = await _basketItemRepository.GetByIdAsync(id);
 
+        if (basketItem is null)
+            throw new BookNotFoundException("The book does not exist in the basket.");
+
         _basketItemRepository.Delete(basketItem);
         await _basketItemRepository.SaveAsync();
 
-        return new ResponseDto((int)HttpStatusCode.OK, "Book is deleted");
+        return new ResponseDto((int)HttpStatusCode.OK, "The book has been deleted from the basket.");
     }
+
+
 }
